+++ pkg/ioutils/readers.go
@@ -2,8 +2,6 @@ package ioutils // import "github.com/docker/docker/pkg/ioutils"
 
 import (
 	"context"
-	"crypto/sha256"
-	"encoding/hex"
 	"io"
 )
 
@@ -47,15 +45,6 @@ func NewReaderErrWrapper(r io.Reader, closer func()) io.Reader {
 		reader: r,
 		closer: closer,
 	}
-}
-
-// HashData returns the sha256 sum of src.
-func HashData(src io.Reader) (string, error) {
-	h := sha256.New()
-	if _, err := io.Copy(h, src); err != nil {
-		return "", err
-	}
-	return "sha256:" + hex.EncodeToString(h.Sum(nil)), nil
 }
 
 // OnEOFReader wraps an io.ReadCloser and a function
